+++ src/index.ts
@@ -20,6 +20,6 @@ export {
 
 export { WakuRelay, RelayCodecs } from './lib/waku_relay';
 
-export { Direction, WakuStore, StoreCodec } from './lib/waku_store';
+export { PageDirection, WakuStore, StoreCodec } from './lib/waku_store';
 
 export * as proto from './proto';